Move kraken staking test
